Html and CSS new classes styles and image dimensions
I made a div for the footer called the -footend the  and in the CSS I madet the body be seemless with a margin of 0, padding of 0 and a background white. Removed the footer classes, gave the hero-class a flex 1 and a max width of 50% so that the heading and the image only take up that amount of the screen at most. I gave the to-action class margins to seperate it from the quote and footer aswell as padding and moves the previous footer class code to the new the-footend style